Feature/fe dev

+++ services/billing-service/src/main/java/com/roomie/services/billing_service/dto/request/BillRequest.java
@@ -28,20 +28,20 @@ public class BillRequest {
     Double waterUnitPrice;
 
     // Internet
-    Double internetPrice;
+    BigDecimal  internetPrice;
 
     // Parking
-    Double parkingPrice;
+    BigDecimal  parkingPrice;
 
     // Cleaning
-    Double cleaningPrice;
+    BigDecimal  cleaningPrice;
 
     // Maintenance
-    Double maintenancePrice;
+    BigDecimal  maintenancePrice;
 
     // Other
     String otherDescription;
-    Double otherPrice;
+    BigDecimal  otherPrice;
 
     String notes;
